+++ src/main/java/io/github/poshjosh/ratelimiter/annotations/RateGroup.java
@@ -1,6 +1,6 @@
 package io.github.poshjosh.ratelimiter.annotations;
 
-import io.github.poshjosh.ratelimiter.util.Operator;
+import io.github.poshjosh.ratelimiter.model.Operator;
 
 import java.lang.annotation.*;