+++ src/test/java/io/github/poshjosh/ratelimiter/performance/dummyclasses/RateGroupA.java
@@ -10,7 +10,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.concurrent.TimeUnit;
 
-@Rate(1)
+@Rate("1/s")
 @Rate(permits = 10, timeUnit = TimeUnit.MINUTES)
 @RateCondition("jvm.memory.available > 100MB")
 @RateGroup("rate-group-A")