+++ src/test/java/io/github/poshjosh/ratelimiter/performance/dummyclasses/RateGroupB.java
@@ -10,7 +10,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.concurrent.TimeUnit;
 
-@Rate(1)
+@Rate("1/s")
 @RateGroup("rate-group-B")
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE, ElementType.METHOD})