+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Method `RateSource#getRates` to get rates defined at the respective source.
 - Nesting of property defined rates, possible via `Rates#parentId`.
 - Method `RateLimiterRegistry#getMatchContext(String id)`.
+- Field `rate` to the `@Rate` annotation. It is the default. Example use: `@Rate("9/s")`.
 
 ### Changed